+++ packages/proto-signing/src/decode.spec.ts
@@ -44,8 +44,9 @@ describe("decode", () => {
                 single: {
                   mode: SignMode.SIGN_MODE_DIRECT,
                 },
+                multi: undefined,
               },
-              sequence: Long.fromNumber(0, true),
+              sequence: Long.UZERO,
             },
           ],
           fee: {