+++ packages/stargate/webpack.web.config.js
@@ -17,10 +17,10 @@ module.exports = [
     },
     plugins: [
       new webpack.EnvironmentPlugin({
-        SIMAPP42_ENABLED: "",
-        SLOW_SIMAPP42_ENABLED: "",
         SIMAPP44_ENABLED: "",
         SLOW_SIMAPP44_ENABLED: "",
+        SIMAPP46_ENABLED: "",
+        SLOW_SIMAPP46_ENABLED: "",
       }),
       new webpack.ProvidePlugin({
         Buffer: ["buffer", "Buffer"],