+++ packages/tendermint-rpc/src/tendermint34/responses.ts
@@ -45,7 +45,9 @@ export interface AbciQueryResponse {
   readonly height?: number;
   readonly index?: number;
   readonly code?: number; // non-falsy for errors
+  readonly codespace: string;
   readonly log?: string;
+  readonly info: string;
 }
 
 export interface BlockResponse {
@@ -192,7 +194,7 @@ export interface Event {
 
 export interface TxData {
   readonly code: number;
-  readonly codeSpace?: string;
+  readonly codespace?: string;
   readonly log?: string;
   readonly data?: Uint8Array;
   readonly events: readonly Event[];
@@ -325,6 +327,9 @@ export interface NodeInfo {
   /** IP and port */
   readonly listenAddr: string;
   readonly network: string;
+  /**
+   * The Tendermint version. Can be empty (see https://github.com/cosmos/cosmos-sdk/issues/7963).
+   */
   readonly version: string;
   readonly channels: string; // ???
   readonly moniker: string;
@@ -347,13 +352,13 @@ export interface SyncInfo {
 export interface Validator {
   readonly address: Uint8Array;
   readonly pubkey?: ValidatorPubkey;
-  readonly votingPower: number;
+  readonly votingPower: bigint;
   readonly proposerPriority?: number;
 }
 
 export interface ValidatorUpdate {
   readonly pubkey: ValidatorPubkey;
-  readonly votingPower: number;
+  readonly votingPower: bigint;
 }
 
 export interface ConsensusParams {